Update module structure, minor doc fixes
