+++ task_geo/data_sources/demographics/us_census/__init__.py
@@ -1,0 +1,3 @@
+from task_geo.data_sources.demographics.us_census import us_census
+
+__all__ = ['us_census']
